requirements.txt
streamlit>=1.52.0
pandas>=2.0.0
openpyxl>=3.1.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
